temporary fix

+++ plugins/image/process/filter/Sobel/src/SobelAlgorithm.hpp
@@ -24,7 +24,8 @@ struct channel_gradientDirection_t
 	GIL_FORCEINLINE
 	void operator()( const Channel& x, const Channel& y, Channel& res ) const
 	{
-		res = std::atan2( y, x );
+		/// @todo tuttle: select floating type
+		res = std::atan2( (double)y, (double)x );
 	}
 };
